feat: php 8.3 support
* feat: Add PHP 8.3 to composer.json

chore: coding-standards

chore: bump ergebnis/php-cs-fixer-config to 6.11.0

fix: PHP-CS-Fixer ruleset creation

chore: bump other packages

* chore: remove unused reference

* fix: Psalm errors

* Apply php-cs-fixer changes

* fix: IntegrationTestCase unlink tmp file on setUp

fix: Workflow actions to use latest versions

* Apply php-cs-fixer changes

* fix: disable PhpCsFixerCustomFixers/phpdoc_array_style breaking psalm check

* chore: include nyholm/psr7 and nyholm/psr7-server to resolve RuntimeException for PSR-17 ResponseFactory implementations

* fix: PHP-CS-Fixer and Psalm

---------

+++ tests/IntegrationTestCase.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace RoachPHP\Tests;
 
 use PHPUnit\Framework\TestCase;
-use Throwable;
 use const JSON_THROW_ON_ERROR;
 
 /**
@@ -28,15 +27,14 @@ abstract class IntegrationTestCase extends TestCase
     {
         $this->skipIfServerNotRunning();
 
-        try {
-            @\unlink(__DIR__ . '/Server/tmp/crawled.json');
-        } catch (Throwable) {
+        if (\file_exists(__DIR__ . '/Server/tmp/crawled.json')) {
+            \unlink(__DIR__ . '/Server/tmp/crawled.json');
         }
     }
 
     protected function skipIfServerNotRunning(): void
     {
-        if (false === @\file_get_contents("{$this->serverUrl}/ping")) {
+        if (false === \file_get_contents("{$this->serverUrl}/ping")) {
             self::markTestSkipped('Skipping integration test. Server not running.');
         }
     }
